Alternative fix for read-only 2.ChangeSet files.
This is a backport of a cset originally created in
bk-dev for handling the case where bk can't append
to a read-only heap file.  It fixes the case where
the 2.ChangeSet file is read-only and owned by
another user.  Instead of trying to fix the permissions
we just rewrite the file.

bk: 5272d858LlqIzVTkBPgFoHoeQRSjoA