Fix a bug in which diffing two large, all different files dumped core.
The bug would manifest as "bk difftool" not printing any diffs.
The real problem was that "bk ndiff --sdiff=1" would dump core
because of the stack allocation in alignMods(). Furthermore,
since the algorithm is quadratic, a large diff block would cause
difftool to hang. Quick stopgap is to not do alignment if we have
diff blocks that are too large. The rationale is that if you're
looking at more than a screenful of diffs, the alignment algorithm
loses value.

bk: 52ced7f27hLuHvSSIOP_yuxHsMdtKA