Make it easier to use syncroot instead of rootkey by keeping it cached in the proj struct. --- Extend BAM data to be stored under the syncroot instead of the rootkey.  This means that components will use the same BAM data when detached and when attached to a different product.
Before, we insisted that detatch take all data with
it.  And so we maintained two separate copies of the
BAM data.

The reason for this RTI is to make bk-archive not lose
data when a component is detached.

For compatibility, a symlink from the old BAM directory name
will be created for versions of bk before this release.

bk: 52cffb08W5JDFW3Ft3nc5QdZtcoa-g