API changes for diff, as used by ndiff and smerge. Diff is now just diffing a range of indexes.
The purpose of this change is to make it easier to diff things,
beyond lines of characters, while hopefully making it simpler
to diff lines of characters.

Core item is a hunk or addArray of hunks.
Internals of hunk have are now arrays (DF_LEFT, DF_RIGHT)
instead of individual elements.

Includes some bugfixes in diff output (see regressions).

bk: 53600bccvVixQxeZigrAXpbF9_6SLQ