Support -S working with -L in command line and gui tools. See that --standalone works as -S in all these commands.
Why: -S is needed for when a nested component is talking with a
standalone repo.  Nice if it could figure that out automaically,
but a least good to have direct support for it.

Not done: now that a number of commands have -S, I would expect
-Sr@ behavior to be different and it isn't, so for now, -S is
illegal without -L in the command line tools.

Rewrite csettool's main() in L to make it easy to handle long args.

bk: 540f61a4wUTo23qyr-UVfQAAZqMR4Q