Workaround a linux permissions problem with lclone
On current ubuntu a link() syscall can fail if the file is owned
by a different user. Our current clone code assumes that if
linking the s.ChangeSet file work then the link() will work
for the rest of the files in the repository. That isn't always
true.  We workaround this by making sfio create a copy of the
link fails.

bk: 5425d114ZVQgX_k7kNwtuczEFcyDkQ