Check performance fixes.
It was observed that most of the time running a partial check was
spent on two tests of the ChangeSet file.  The tip checksum
validations and a pathname test that looks for path conflicts in
the deltakeys of the tip cset.

The performance of the path conflict code was doubled by avoiding
copying data out of the heap that wasn't needed.  rdweavePair was
used to not read data not in the tip and offsets to the heap
where used instead of strings.

The checksum code performance was also doubled by making use
of the new rset code.

Rick contributed heavily to this cset.

bk: 5440f077PY0TXxxZNJjmlJ_3vcvirg