The short circuit code was too agressive in the face of wacky bk graphs.  See test case in t.rset for example.
It is possible using gone to have a rootkey subgraph have
multiple tips, such that the newest is both gone and shows
up in bk get ChangeSet.

If the first instance of a node was seen in the gca, then
it was ignored, since gca means all other nodes would
descend from this and need to be gca... except in the
case mentioned where there's some other unrelated node.

The short circuit code was simplified / weakened to
put more nodes in the hash to block seeing the 
(rare) other tip.  This result is to, in some cases,
bloat the hash, and sometimes read more of the weave
(as shown in the regression).

The prune code in sortKey was regularized (handle left1
and left2 the same).

bk: 54885636dunWLIjVgoMz7z2iQoqEGA