Fix a problem Larry discovered.
Some older versions of bk could have a heap where a some strings that should
be collapsed together are duplicated.  The current bk shouldn't do this and
we added an assertion to say this shouldn't happen, but that assertion
can trip on repos created with old bk.

Just remove the assertion.

bk: 54b01b11XwXkCIQ0d3EHPb_35mLSdw