Various uniqdb fixes.
Fix bug where a ChangeSet delta key that was written to
the uniqdb would not be read back correctly.

Refactor the uniqdb request code. The main change is to handle the
connect/select/send/recv errors differently. When moving from TCP to
UDP, the failure modes changed but the UDP code didn't reflect that.
When we used TCP, we could tell immediately if the uniq_daemon was no
longer listening on a particular port because the connect or send
would fail immediately. In contrast, the nature of UDP is
connectionless, unreliable delivery, so there is no connect per-se and
the send does not seem to fail if the daemon isn't listening.  Only
the recv fails and only after the timeout period, which was 10
seconds. I observed this on OS X and Linux, anyway.

In practice, the uniq_daemon isn't started and stopped that much, so
suffering a timeout on the order of a second or seconds every now and
then is probably OK. But regressions tend to stress the startup and
shutdown cases more and trip more recv timeouts. I arbitrarily reduced
the timeout to 1 second and added backoff logic to increase it after
each retry.

Cd to and stay in /tmp while running instead of sitting in the
db dir. We were having a problem with the daemon holding the
dir open in its exit path.

Leave in code to arm retry-tracing debug if _BK_UNIQ_TRACE_THRESH
is set.

Some riders:

On ppc we kept getting stuff left over in t.setup.
Two reasons:
  - t.setup cleared BK_REGRESSIONS as we ran _lconfig
  - t.setup changed BK_DOTBK and we didn't send --quit to
    the uniqdb
(Wayne)

Win8 was hung in _popensystem and that wasn't made for windows.
(Wayne)

Before running a test, set BK_CURRENT_TEST in the environment.
We do this so that using Process Explorer on Windows we can figure
out which script is running by looking at the environment (otherwise
we can only see that sh.exe is running).
(Wayne)

Fix missing newline in t.L test.
(Rob)

bk: 54e9c317fcD_8HsGUYElvmFDfPd5Hg