Pull -rTAGKEY fails in synckeys
Bluearc found that pull could assert in probekey if we pass a tagkey
in the -r arguement. This is because the synckey code assumes it is
dealing with csetkeys and sends all implied tags automatically.

Push had the same bug.

(includes bk-6.1.3 release notes)

bk: 54f0e810YeSRH40AFYID3ba5aUAwig