Close a uniqdb loophole that could prevent the uniq_server from starting if it ever leaves a stale port file behind. This can happen if the uniq_server crashes or is killed (Oscar hit this in his testing after a kill -9 of his bk processes).
We used to have logic that checked for a stale	or missing lock file
before trusting the port file (in uniqdb_socket()), but that did the
check on every call to uniq_adjust() for what is a relatively rare
case. Instead of adding that back in, the fix here doesn't slow down
the common case but will still recover from a stale or missing lock
file.

Also, Wayne noticed that we weren't closing an fmem.

bk: 54f4fc174IYA66Q6LxVx-R2ck-hfAA