Perf work for abnormal case; slight perf degrade for normal case.
Oscar's import was pulling 50k csets in a patch.  That was
pushing on the fmem system to have 50k FILE *handles open.
Stdio has a O(N) degradation in fopen.
  
This fix uses DATA block instead of fmem to hold the data
in the patch list.  Then when processing, it puts it back
into a FILE* with fmem_buf().
  
Long term, an intermediate stage of takepatch goes away, and
along with it, all the changes of this RTI.  This is just
a short term fix to the huge (in number of csets) pull case.

bk: 54f5afe83G_osXmaJZAR2YjAHKkbJg