Fix a paging corruption issue that appears on some NFS machines. This also happens to be a performance fix.
A simple typo in sccs.h caused the CHKXOR feature of fopen_crc()
to be enabled in more cases than intended. This causes us to reread
s.ChangeSet and [12].ChangeSet on fclose() to fully verify checksums.
(This is intended to only be enabled by check -ac)

Then in bin_writeHeap() we have a code sequence where new data added to
the heap as part of the current operation is appended to the 2.ChangeSet
file.  Meanwhile if the heap data is large enough then paging is enabled
as we have a s->heapfh[2] FILE* open to the 2.ChangeSet file that is
used for paging in data to the s->heap.buf array. Normally s->heapfh[2]
will never read the new portion of the file written by bin_writeHeap()
and so it works fine.  But when CHKXOR was incorrectly enabled, in
sccs_free() we call fclose(s->heapfh[2]) and that triggers the checking
code to reread the entire file using the EOF marker to know when we have
reached the end.  What we are seeing is that since we had already read
the block right before the XOR block from before the file was extended
then we reuse that old data and incorrectly determine that the file ends
early without reading to the end of the file.  However since we had not
previously read the xor block when that block is read we get part of
the new data that does pass its own crc32 it doesn't contain XOR data
and so doesn't validate.

bk: 551d1b696FAn1HNuudsy-Sy3fTADLw