Oscar found a problem in partition that can create poly without the poly db.
The fix is to drop the check, allowing this case.

I looked into creating the polydb entry as part of partition,
and it is not easy, because prior product cset keys go into
the poly db, which influences the corresponding delta key
which goes into a product merge cset.  Lots of shuffling.

bk: 552e63072TY7sbVmE-mrtXlE7ktdqA