Don't flush leases or remove accepted files any more, and don't pop up any screens in installtool if there's already a license in the config and the EULA for it has been accepted (according to bk _eula -u). The net effect is that, on macos, you will see the EULA iff bk _eula -u says you haven't seen and accepted it yet, and you no longer will be prompted to accept the EULA on the first bk command you run after installing.
Sudo as the user when running installtool and other bk commands that
acquire a lease, instead of letting it run as root. This avoids crumbs
left behind in /tmp.

Installtool bug fix for not being able to choose any file in the
license dialog box on OS X. If you wanted to use a license file
that did not end in .txt or .lic, a Tk bug was keeping it greyed
out in the file chooser and not letting you select it.

Rider: installtool bugfix from Damon for when a license file
is chosen with the license stuff right at the top. I had been
testing with my config file, where the license and licsign
lines were further down and that didn't trip the bug.

bk: 555f4d67_wpKlLIsnbOaN14p0LyQ-A