Use ssh for copying the images from windows
Rsh on Windows is flaky. It needs .rhosts just right and even
with the proper configuration, I've seen it fail to copy
the image to work. Ssh is better maintained by the cygwin
people so revert to that. This means that 'work' has to have
the proper configuration: 1) an 'image' account set up 
with the private keys in .ssh/authorized_hosts, and 2) the
right permissions so the user 'image' can write to /home/bk/images.

bk: 5570cb0aNs-JabclfnHgT5vtITmxXg