Fix a performance bug in bk diffs.
Larry found out we were sccs_get()'ing the ChangeSet file for 
each modified file when running "bk -Uc diffs". The bug turned
out to be a combination of mapRev() changing the NULL for rrev
into "edited" for the right revision and later getHistoricPath()
calling sccs_findrev().

sccs_findrev() can't tell what "edited" means so it looks it up
as a tag in the ChangeSet file, thus producing the behavior that
Larry was seeing. 

The fix is to call findrev() which doesn't handle tags. It's okay
to call it here because mapRev() has already called sccs_findrev()
for both sides, so all we need here is findrev().

bk: 5570e1dcfj6xv69g06KUVguZlzqP-g