Expose errors if the resolver fails to remove a directory.
Rich recently hit a bug where a directory was replaced with
a symlink and the resolver asked him what to do. He chose 
'rl' to remove the local directory, but there was an ignored
file in there so the rmdir() failed. Since it fails silently, 
the resolver looks like it went in an infinite loop. This cset
prints the reason for the failure and the list of files found
so that if the file blocking the rmdir() starts with dot, 
(e.g. ".ignored-file") it's not hard to find them.

bk: 559aa8083H84oYPc06km14zqMyfBng