Prevent 'bk names' from leaving the repository in a mess after an error.
This is a cset from 2005.  Save a log of renames attempted and when an error
condition is encountered, rewind those renames to put everything back
in its original location.

Also fix the code to mantain the idcache correctly and handle a couple more
corner cases correctly.

bk: 55fb107fcB7joKLmq8rSiTWXR2Lj3g