Handle repos in the wild that have excludes of merge csets, which resulted in improper bookkeeping leading to a mismatch in the ChangeSet file generated and the file tips after a clone -r.
This shows up as a dual sum failure (one sum being serialmap based
and the other being walkrevs based).

Instead of fixing the repos so checksums match, this fix ignores
a region of history where the failure happened.  The configuration
can be done by the user with no_graphverify, or it can be baked into bk.

bk: 5643797dNWrNUxua0b88Gaqgpb2qag