Update 'bk makepatch' so it ONLY generates the bk-5.0 fastpatch format. And in fact defaults to the bk-7.0 BKMERGE format.
We now generate the fast patch format with or without the -F option to
'bk makepatch'.  This breaks pull/push compatibility with bk-4.x, but
old repositories can still be cloned and read. Also we no longer put
sfiles in a SFIO as part of a patch. This isn't really needed for
performance anymore and has been a source of bugs in the past.

bk: 56ce0e33FUIJpVbpOqQ4HnP18OTwnw