Add bk changes [--lattice|--longest].  Remember to use with -e, or some nodes in the range won't be listed.
Useful when either wanting an arbitrary line from
root to tip, where choice is influenced by length.
Or wanting a form of bisect where the older bound is
always present in the csets being tested.
The normal range A..B can yield csets where A is
not reachable in their history.

bk: 56d5bc676Va5P973TRAraOB2gTUn-A