Change the --compat (oldest support repository format) for the OSS release to be bk-7.0.
Here I attempt to make it harder for bk from the command line
to generate older non-BKMERGE repositories so reduce support.

Lots of help from Rick to keep testcases in a reasonable shape with this
change and to document what the tests are actually doing.

bk: 56e6bd20DBXXQS_uX2xJI5Y-wAxd1Q