Replace the tag walking code with a faster version.
Extracted from the tag-deletes RTI, this adds a with sccs_walkTag() routine to
find tags pointing at a given cset.  It caches the tag structure in 'sccs*' so
this is O(1) instead of O(n) (or O(n^2) for tag search)

Also includes a greatly expanded tag graph validation routine called from check
that was used to understand how the current code works.

bk: 56facfbfKKCXUYtVwGffvVRgfa1Gag