Cleanups of directory browsing in bkweb.
Now when browsing repositories, we skip and files
starting with '.'.  (like .ssh )

When browsing inside the repository, we no longer look at
files on the disk, but instead use rset to browse the committed
history.  This hides any other files in the repository, but also
enables browsing of the past.
(This is a cool experimental feature at the moment)

bk: 572ce1e3DSFGabtqZ-IlECC0zX_mZw