Make fast-export actually fast.
While exporting the 1M tree I noticed that we were exporting the
file contents at around 100 MB/s but once we started importing
changesets we dropped to 4KB/s with long pauses.

In that repo, rset_diff() was taking around 120 ms, which translates
to about 8 csets/sec. So my ETA for exporting the 1M repo was around 
34 hours.

With this code it takes 7 minutes.

bk: 57487fbbw2tAMWB1ns2m_OMLbDAkYA