Add ability to delete tags.
This is a fairly large cset that touches a lot of areas.
BitKeeper has always supported revision controlled tags and you could move tags to new
csets.  In the past it wasn't possible to see that this happened with 'bk changes'.
Both locations would show the tags and the :TAG: dspec would return the same information
for both locations.  Only running 'bk changes -rTAG' would show the truth.

With this change we add a :TAGS: dspec that annotates the tag information to indicate if
a tag has moved and a :TAGGED: dspec that only returns tags that actually live on a 
given delta at the top of the current repository.

Also in this cset the $each() dspec function has be changed to strictly split a field into
lines and process each line.  The old form was very inconsistent.

bk: 57756de0jWMvTQHlbaZSsh5FBeyfPg