Fix strange build dependancies by not using 'bk' to build 'bk'
Now the 'bk' executable can be built without using bk because I moved
stuff like the 'bk describe' output (that used to be compiled in) to
a file in the install directory.

So now the build was changed to build bk first and then use that bk
for other places that are needed for a clean build.

As normal we were not able to completely do this for Windows so
Windows still requires bk to be installed first for a clean build.

bk: 5776da6f2FZNN-uKBFPZC6VU695tsw